examples/blockchainnews-agent/requirements.txt
bnbagent
ddgs>=0.0.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
